Refactor CustomButtonCalculatorWidget to use Padding and MaterialButton
Replace the Container widget with Padding and MaterialButton in the CustomButtonCalculatorWidget. This change improves the widget's structure and aligns it more closely with Flutter's best practices for button widgets. The MaterialButton now includes a rounded rectangle shape with a border radius of 10. The button's color is set to ColorManager.greenColor, and it spans the full width of its parent container.